+++ funcionario.h
@@ -1,0 +1,5 @@
+void tela_menu_funcionarios();
+    void cadast_funcionario(void);
+    void exibir_funcionario(void);
+    void edit_funcionario(void);
+    void delet_funcionario(void);
