+++ gestao.h
@@ -1,0 +1,5 @@
+void tela_menu_gestao_produtos();
+    void cadast_produto(void);
+    void exibir_produto(void);
+    void edit_produto(void);
+    void delet_produto(void);
